refactor(tools): extract memory and jobs domains from tools.rs (#1613)
