+++ crates/tau-tools/src/tools.rs
@@ -152,10 +152,14 @@ const BUILTIN_AGENT_TOOL_NAMES: &[&str] = &[
     "bash",
 ];
 
+mod jobs_tools;
+mod memory_tools;
 mod registry_core;
 mod runtime_helpers;
 mod session_tools;
 
+pub use jobs_tools::{JobsCancelTool, JobsCreateTool, JobsListTool, JobsStatusTool};
+pub use memory_tools::{MemoryReadTool, MemorySearchTool, MemoryTreeTool, MemoryWriteTool};
 use registry_core::BashSandboxSpec;
 pub use registry_core::{
     builtin_agent_tool_names, register_builtin_tools, register_extension_tools,
@@ -810,1311 +814,6 @@ impl AgentTool for EditTool {
         ToolExecutionResult::ok(json!({
             "path": resolved.display().to_string(),
             "replacements": replacements,
-        }))
-    }
-}
-
-/// Public struct `MemoryWriteTool` used across Tau components.
-pub struct MemoryWriteTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl MemoryWriteTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for MemoryWriteTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "memory_write".to_string(),
-            description:
-                "Write or update a scoped semantic memory entry in the runtime memory store"
-                    .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "memory_id": {
-                        "type": "string",
-                        "description": "Optional stable memory id. A deterministic id is generated when omitted."
-                    },
-                    "summary": {
-                        "type": "string",
-                        "description": format!(
-                            "Memory summary text (max {} characters)",
-                            self.policy.memory_write_max_summary_chars
-                        )
-                    },
-                    "tags": {
-                        "type": "array",
-                        "items": { "type": "string" },
-                        "description": format!(
-                            "Optional tags (max {} items, {} chars each)",
-                            self.policy.memory_write_max_tags,
-                            self.policy.memory_write_max_tag_chars
-                        )
-                    },
-                    "facts": {
-                        "type": "array",
-                        "items": { "type": "string" },
-                        "description": format!(
-                            "Optional facts (max {} items, {} chars each)",
-                            self.policy.memory_write_max_facts,
-                            self.policy.memory_write_max_fact_chars
-                        )
-                    },
-                    "workspace_id": { "type": "string" },
-                    "channel_id": { "type": "string" },
-                    "actor_id": { "type": "string" },
-                    "source_event_key": { "type": "string" },
-                    "recency_weight_bps": {
-                        "type": "integer",
-                        "description": "Optional recency weight in basis points (0..=10000)"
-                    },
-                    "confidence_bps": {
-                        "type": "integer",
-                        "description": "Optional confidence score in basis points (0..=10000)"
-                    }
-                },
-                "required": ["summary"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        let summary = match required_string(&arguments, "summary") {
-            Ok(summary) => summary.trim().to_string(),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_write",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if summary.is_empty() {
-            return ToolExecutionResult::error(json!({
-                "tool": "memory_write",
-                "reason_code": "memory_empty_summary",
-                "error": "summary must not be empty",
-            }));
-        }
-        if summary.chars().count() > self.policy.memory_write_max_summary_chars {
-            return ToolExecutionResult::error(json!({
-                "tool": "memory_write",
-                "reason_code": "memory_summary_too_large",
-                "max_summary_chars": self.policy.memory_write_max_summary_chars,
-                "error": format!(
-                    "summary exceeds max length of {} characters",
-                    self.policy.memory_write_max_summary_chars
-                ),
-            }));
-        }
-
-        let memory_id = arguments
-            .get("memory_id")
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_string)
-            .unwrap_or_else(generate_memory_id);
-
-        let tags = match optional_string_array(
-            &arguments,
-            "tags",
-            self.policy.memory_write_max_tags,
-            self.policy.memory_write_max_tag_chars,
-        ) {
-            Ok(tags) => tags,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_write",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let facts = match optional_string_array(
-            &arguments,
-            "facts",
-            self.policy.memory_write_max_facts,
-            self.policy.memory_write_max_fact_chars,
-        ) {
-            Ok(facts) => facts,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_write",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let recency_weight_bps = match optional_basis_points(&arguments, "recency_weight_bps") {
-            Ok(value) => value.unwrap_or(0),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_write",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let confidence_bps = match optional_basis_points(&arguments, "confidence_bps") {
-            Ok(value) => value.unwrap_or(0),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_write",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let scope = MemoryScope {
-            workspace_id: arguments
-                .get("workspace_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default()
-                .to_string(),
-            channel_id: arguments
-                .get("channel_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default()
-                .to_string(),
-            actor_id: arguments
-                .get("actor_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default()
-                .to_string(),
-        };
-        let source_event_key = arguments
-            .get("source_event_key")
-            .and_then(Value::as_str)
-            .unwrap_or_default()
-            .to_string();
-        let entry = MemoryEntry {
-            memory_id: memory_id.clone(),
-            summary: summary.clone(),
-            tags,
-            facts,
-            source_event_key,
-            recency_weight_bps,
-            confidence_bps,
-        };
-        let store = FileMemoryStore::new_with_embedding_provider(
-            self.policy.memory_state_dir.clone(),
-            self.policy.memory_embedding_provider_config(),
-        );
-        let storage_path = store
-            .storage_path()
-            .map(Path::to_path_buf)
-            .unwrap_or_else(|| store.root_dir().join("memory.backend"));
-
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "memory_write",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({
-                "memory_id": memory_id.clone(),
-                "scope": {
-                    "workspace_id": scope.workspace_id.clone(),
-                    "channel_id": scope.channel_id.clone(),
-                    "actor_id": scope.actor_id.clone(),
-                },
-                "summary_chars": summary.chars().count(),
-                "tags": entry.tags.clone(),
-                "facts": entry.facts.clone(),
-            }),
-        ) {
-            return rbac_result;
-        }
-
-        if let Some(approval_result) = evaluate_tool_approval_gate(ApprovalAction::ToolWrite {
-            path: storage_path.display().to_string(),
-            content_bytes: summary.len(),
-        }) {
-            return approval_result;
-        }
-
-        if let Some(rate_limit_result) = evaluate_tool_rate_limit_gate(
-            &self.policy,
-            "memory_write",
-            json!({
-                "scope": {
-                    "workspace_id": scope.workspace_id.clone(),
-                    "channel_id": scope.channel_id.clone(),
-                    "actor_id": scope.actor_id.clone(),
-                },
-                "summary_chars": summary.chars().count(),
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-            }),
-        ) {
-            return rate_limit_result;
-        }
-
-        match store.write_entry(&scope, entry) {
-            Ok(result) => ToolExecutionResult::ok(json!({
-                "tool": "memory_write",
-                "created": result.created,
-                "memory_id": result.record.entry.memory_id,
-                "scope": result.record.scope,
-                "summary": result.record.entry.summary,
-                "tags": result.record.entry.tags,
-                "facts": result.record.entry.facts,
-                "source_event_key": result.record.entry.source_event_key,
-                "recency_weight_bps": result.record.entry.recency_weight_bps,
-                "confidence_bps": result.record.entry.confidence_bps,
-                "updated_unix_ms": result.record.updated_unix_ms,
-                "embedding_source": result.record.embedding_source,
-                "embedding_model": result.record.embedding_model,
-                "embedding_reason_code": result.record.embedding_reason_code,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-            })),
-            Err(error) => ToolExecutionResult::error(json!({
-                "tool": "memory_write",
-                "reason_code": "memory_backend_error",
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-                "error": error.to_string(),
-            })),
-        }
-    }
-}
-
-/// Public struct `MemoryReadTool` used across Tau components.
-pub struct MemoryReadTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl MemoryReadTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for MemoryReadTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "memory_read".to_string(),
-            description: "Read a scoped semantic memory entry by id".to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "memory_id": { "type": "string", "description": "Memory id to load" },
-                    "workspace_id": { "type": "string" },
-                    "channel_id": { "type": "string" },
-                    "actor_id": { "type": "string" }
-                },
-                "required": ["memory_id"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        let memory_id = match required_string(&arguments, "memory_id") {
-            Ok(memory_id) => memory_id,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_read",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let scope_filter = memory_scope_filter_from_arguments(&arguments);
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "memory_read",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({
-                "memory_id": memory_id.clone(),
-                "scope_filter": scope_filter.clone(),
-            }),
-        ) {
-            return rbac_result;
-        }
-
-        let store = FileMemoryStore::new_with_embedding_provider(
-            self.policy.memory_state_dir.clone(),
-            self.policy.memory_embedding_provider_config(),
-        );
-
-        if let Some(rate_limit_result) = evaluate_tool_rate_limit_gate(
-            &self.policy,
-            "memory_read",
-            json!({
-                "memory_id": memory_id.clone(),
-                "scope_filter": scope_filter.clone(),
-                "store_root": self.policy.memory_state_dir.display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-            }),
-        ) {
-            return rate_limit_result;
-        }
-        match store.read_entry(memory_id.as_str(), scope_filter.as_ref()) {
-            Ok(Some(record)) => ToolExecutionResult::ok(json!({
-                "tool": "memory_read",
-                "found": true,
-                "memory_id": record.entry.memory_id,
-                "scope": record.scope,
-                "summary": record.entry.summary,
-                "tags": record.entry.tags,
-                "facts": record.entry.facts,
-                "source_event_key": record.entry.source_event_key,
-                "recency_weight_bps": record.entry.recency_weight_bps,
-                "confidence_bps": record.entry.confidence_bps,
-                "updated_unix_ms": record.updated_unix_ms,
-                "embedding_source": record.embedding_source,
-                "embedding_model": record.embedding_model,
-                "embedding_reason_code": record.embedding_reason_code,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-            })),
-            Ok(None) => ToolExecutionResult::ok(json!({
-                "tool": "memory_read",
-                "found": false,
-                "memory_id": memory_id,
-                "scope_filter": scope_filter.clone(),
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-            })),
-            Err(error) => ToolExecutionResult::error(json!({
-                "tool": "memory_read",
-                "reason_code": "memory_backend_error",
-                "memory_id": memory_id,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-                "error": error.to_string(),
-            })),
-        }
-    }
-}
-
-/// Public struct `MemorySearchTool` used across Tau components.
-pub struct MemorySearchTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl MemorySearchTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for MemorySearchTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "memory_search".to_string(),
-            description: "Search semantic memory entries using deterministic similarity ranking"
-                .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string", "description": "Semantic search query" },
-                    "workspace_id": { "type": "string" },
-                    "channel_id": { "type": "string" },
-                    "actor_id": { "type": "string" },
-                    "limit": {
-                        "type": "integer",
-                        "description": format!(
-                            "Maximum matches to return (default {}, max {})",
-                            self.policy.memory_search_default_limit,
-                            self.policy.memory_search_max_limit
-                        )
-                    }
-                },
-                "required": ["query"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        let query = match required_string(&arguments, "query") {
-            Ok(query) => query,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_search",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if query.trim().is_empty() {
-            return ToolExecutionResult::error(json!({
-                "tool": "memory_search",
-                "reason_code": "memory_empty_query",
-                "error": "query must not be empty",
-            }));
-        }
-        let limit = match optional_usize(
-            &arguments,
-            "limit",
-            self.policy.memory_search_default_limit,
-            self.policy
-                .memory_search_max_limit
-                .max(self.policy.memory_search_default_limit),
-        ) {
-            Ok(limit) => limit,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "memory_search",
-                    "reason_code": "memory_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let scope = memory_scope_filter_from_arguments(&arguments).unwrap_or_default();
-
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "memory_search",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({
-                "query": query.clone(),
-                "limit": limit,
-                "scope_filter": scope.clone(),
-            }),
-        ) {
-            return rbac_result;
-        }
-
-        let store = FileMemoryStore::new_with_embedding_provider(
-            self.policy.memory_state_dir.clone(),
-            self.policy.memory_embedding_provider_config(),
-        );
-
-        if let Some(rate_limit_result) = evaluate_tool_rate_limit_gate(
-            &self.policy,
-            "memory_search",
-            json!({
-                "query": query.clone(),
-                "limit": limit,
-                "scope_filter": scope.clone(),
-                "store_root": self.policy.memory_state_dir.display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-            }),
-        ) {
-            return rate_limit_result;
-        }
-        match store.search(
-            query.as_str(),
-            &MemorySearchOptions {
-                scope,
-                limit,
-                embedding_dimensions: self.policy.memory_embedding_dimensions,
-                min_similarity: self.policy.memory_min_similarity,
-                enable_hybrid_retrieval: self.policy.memory_enable_hybrid_retrieval,
-                bm25_k1: self.policy.memory_bm25_k1,
-                bm25_b: self.policy.memory_bm25_b,
-                bm25_min_score: self.policy.memory_bm25_min_score,
-                rrf_k: self.policy.memory_rrf_k,
-                rrf_vector_weight: self.policy.memory_rrf_vector_weight,
-                rrf_lexical_weight: self.policy.memory_rrf_lexical_weight,
-                enable_embedding_migration: self.policy.memory_enable_embedding_migration,
-                benchmark_against_hash: self.policy.memory_benchmark_against_hash,
-                benchmark_against_vector_only: self.policy.memory_benchmark_against_vector_only,
-            },
-        ) {
-            Ok(result) => ToolExecutionResult::ok(json!({
-                "tool": "memory_search",
-                "query": result.query,
-                "limit": limit,
-                "scanned": result.scanned,
-                "returned": result.returned,
-                "retrieval_backend": result.retrieval_backend,
-                "retrieval_reason_code": result.retrieval_reason_code,
-                "min_similarity": self.policy.memory_min_similarity,
-                "embedding_dimensions": self.policy.memory_embedding_dimensions,
-                "embedding_backend": result.embedding_backend,
-                "embedding_reason_code": result.embedding_reason_code,
-                "migrated_entries": result.migrated_entries,
-                "query_embedding_latency_ms": result.query_embedding_latency_ms,
-                "ranking_latency_ms": result.ranking_latency_ms,
-                "lexical_ranking_latency_ms": result.lexical_ranking_latency_ms,
-                "fusion_latency_ms": result.fusion_latency_ms,
-                "vector_candidates": result.vector_candidates,
-                "lexical_candidates": result.lexical_candidates,
-                "baseline_hash_overlap": result.baseline_hash_overlap,
-                "baseline_vector_overlap": result.baseline_vector_overlap,
-                "matches": result.matches,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-            })),
-            Err(error) => ToolExecutionResult::error(json!({
-                "tool": "memory_search",
-                "reason_code": "memory_backend_error",
-                "query": query,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-                "error": error.to_string(),
-            })),
-        }
-    }
-}
-
-/// Public struct `MemoryTreeTool` used across Tau components.
-pub struct MemoryTreeTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl MemoryTreeTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for MemoryTreeTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "memory_tree".to_string(),
-            description: "Render memory store hierarchy (workspace -> channel -> actor)"
-                .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {},
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, _arguments: Value) -> ToolExecutionResult {
-        let store = FileMemoryStore::new_with_embedding_provider(
-            self.policy.memory_state_dir.clone(),
-            self.policy.memory_embedding_provider_config(),
-        );
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "memory_tree",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({
-                "store_root": self.policy.memory_state_dir.display().to_string(),
-            }),
-        ) {
-            return rbac_result;
-        }
-        if let Some(rate_limit_result) = evaluate_tool_rate_limit_gate(
-            &self.policy,
-            "memory_tree",
-            json!({
-                "store_root": self.policy.memory_state_dir.display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-            }),
-        ) {
-            return rate_limit_result;
-        }
-        match store.tree() {
-            Ok(tree) => ToolExecutionResult::ok(json!({
-                "tool": "memory_tree",
-                "total_entries": tree.total_entries,
-                "workspaces": tree.workspaces,
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-            })),
-            Err(error) => ToolExecutionResult::error(json!({
-                "tool": "memory_tree",
-                "reason_code": "memory_backend_error",
-                "store_root": store.root_dir().display().to_string(),
-                "storage_backend": store.storage_backend_label(),
-                "backend_reason_code": store.storage_backend_reason_code(),
-                "storage_path": store
-                    .storage_path()
-                    .map(|path| path.display().to_string()),
-                "error": error.to_string(),
-            })),
-        }
-    }
-}
-
-/// Public struct `JobsCreateTool` used across Tau components.
-pub struct JobsCreateTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl JobsCreateTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for JobsCreateTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "jobs_create".to_string(),
-            description:
-                "Queue a background job command for asynchronous execution and persisted tracking"
-                    .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "command": {
-                        "type": "string",
-                        "description": "Executable command (e.g. cargo, sh, cmd)"
-                    },
-                    "args": {
-                        "type": "array",
-                        "items": { "type": "string" },
-                        "description": "Optional command arguments"
-                    },
-                    "cwd": {
-                        "type": "string",
-                        "description": "Optional working directory constrained to allowed roots"
-                    },
-                    "timeout_ms": {
-                        "type": "integer",
-                        "description": format!(
-                            "Optional timeout in milliseconds (1..={})",
-                            self.policy.jobs_max_timeout_ms
-                        )
-                    },
-                    "env": {
-                        "type": "object",
-                        "description": "Optional environment map. Values must be strings.",
-                        "additionalProperties": { "type": "string" }
-                    },
-                    "channel_transport": {
-                        "type": "string",
-                        "description": "Optional channel transport for channel-store trace emission"
-                    },
-                    "channel_id": {
-                        "type": "string",
-                        "description": "Optional channel id for channel-store trace emission"
-                    },
-                    "session_path": {
-                        "type": "string",
-                        "description": "Optional session path override for trace message append"
-                    }
-                },
-                "required": ["command"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        if !self.policy.jobs_enabled {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_create",
-                "reason_code": "jobs_disabled",
-                "error": "background jobs tooling is disabled by policy",
-            }));
-        }
-
-        let command = match required_string(&arguments, "command") {
-            Ok(command) => command.trim().to_string(),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if command.is_empty() {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_create",
-                "reason_code": "jobs_invalid_arguments",
-                "error": "command must not be empty",
-            }));
-        }
-
-        let args = match optional_string_array_unbounded(&arguments, "args") {
-            Ok(args) => args,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let timeout_ms = match optional_positive_u64(&arguments, "timeout_ms") {
-            Ok(timeout_ms) => timeout_ms,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        if let Some(timeout_ms) = timeout_ms {
-            if timeout_ms > self.policy.jobs_max_timeout_ms {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": format!(
-                        "timeout_ms exceeds policy maximum {}",
-                        self.policy.jobs_max_timeout_ms
-                    ),
-                }));
-            }
-        }
-
-        let env = match optional_string_map(&arguments, "env") {
-            Ok(env) => env,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let cwd = match optional_string(&arguments, "cwd") {
-            Ok(cwd) => cwd,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let cwd = match cwd {
-            Some(cwd) => match resolve_and_validate_path(&cwd, &self.policy, PathMode::Read) {
-                Ok(cwd) => Some(cwd),
-                Err(error) => {
-                    return ToolExecutionResult::error(json!({
-                        "tool": "jobs_create",
-                        "reason_code": "jobs_invalid_arguments",
-                        "error": error,
-                    }))
-                }
-            },
-            None => None,
-        };
-
-        let session_path_override = match optional_string(&arguments, "session_path") {
-            Ok(path) => path,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let session_path = match session_path_override {
-            Some(path) => match resolve_and_validate_path(&path, &self.policy, PathMode::Write) {
-                Ok(path) => Some(path),
-                Err(error) => {
-                    return ToolExecutionResult::error(json!({
-                        "tool": "jobs_create",
-                        "reason_code": "jobs_invalid_arguments",
-                        "error": error,
-                    }))
-                }
-            },
-            None => self.policy.jobs_default_session_path.clone(),
-        };
-
-        let channel_transport = match optional_string(&arguments, "channel_transport") {
-            Ok(value) => value,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let channel_id = match optional_string(&arguments, "channel_id") {
-            Ok(value) => value,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if (channel_transport.is_some() && channel_id.is_none())
-            || (channel_transport.is_none() && channel_id.is_some())
-        {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_create",
-                "reason_code": "jobs_invalid_arguments",
-                "error": "channel_transport and channel_id must both be provided when tracing to channel-store",
-            }));
-        }
-
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "jobs_create",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({
-                "command": command.clone(),
-                "args_count": args.len(),
-                "cwd": cwd.as_ref().map(|path| path.display().to_string()),
-                "timeout_ms": timeout_ms.unwrap_or(self.policy.jobs_default_timeout_ms),
-            }),
-        ) {
-            return rbac_result;
-        }
-        if let Some(approval_result) = evaluate_tool_approval_gate(ApprovalAction::ToolBash {
-            command: command.clone(),
-            cwd: cwd.as_ref().map(|path| path.display().to_string()),
-        }) {
-            return approval_result;
-        }
-        if let Some(rate_limit_result) = evaluate_tool_rate_limit_gate(
-            &self.policy,
-            "jobs_create",
-            json!({
-                "command": command.clone(),
-                "args_count": args.len(),
-                "timeout_ms": timeout_ms.unwrap_or(self.policy.jobs_default_timeout_ms),
-            }),
-        ) {
-            return rate_limit_result;
-        }
-
-        let runtime = match resolve_background_job_runtime(&self.policy) {
-            Ok(runtime) => runtime,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_runtime_unavailable",
-                    "error": error,
-                }))
-            }
-        };
-
-        let request = BackgroundJobCreateRequest {
-            command,
-            args,
-            env,
-            cwd,
-            timeout_ms,
-            trace: BackgroundJobTraceContext {
-                channel_store_root: Some(self.policy.jobs_channel_store_root.clone()),
-                channel_transport,
-                channel_id,
-                session_path,
-            },
-        };
-        let record = match runtime.create_job(request).await {
-            Ok(record) => record,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_create",
-                    "reason_code": "jobs_runtime_error",
-                    "error": error.to_string(),
-                }))
-            }
-        };
-        let health = runtime.inspect_health().await;
-
-        ToolExecutionResult::ok(json!({
-            "tool": "jobs_create",
-            "reason_code": "job_queued",
-            "job": background_job_record_payload(&record, false),
-            "health": background_job_health_payload(&health),
-        }))
-    }
-}
-
-/// Public struct `JobsListTool` used across Tau components.
-pub struct JobsListTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl JobsListTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for JobsListTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "jobs_list".to_string(),
-            description:
-                "List persisted background jobs with optional status filtering and runtime health"
-                    .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "limit": {
-                        "type": "integer",
-                        "description": format!(
-                            "Max jobs returned (default {}, max {})",
-                            self.policy.jobs_list_default_limit,
-                            self.policy.jobs_list_max_limit
-                        )
-                    },
-                    "status": {
-                        "type": "string",
-                        "description": "Optional status filter: queued, running, succeeded, failed, cancelled, terminal"
-                    }
-                },
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        if !self.policy.jobs_enabled {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_list",
-                "reason_code": "jobs_disabled",
-                "error": "background jobs tooling is disabled by policy",
-            }));
-        }
-
-        let limit = match optional_usize(
-            &arguments,
-            "limit",
-            self.policy.jobs_list_default_limit,
-            self.policy
-                .jobs_list_max_limit
-                .max(self.policy.jobs_list_default_limit),
-        ) {
-            Ok(limit) => limit,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_list",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let status = match optional_string(&arguments, "status") {
-            Ok(status) => status,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_list",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        let status_filter = match status.as_deref() {
-            Some(raw) => match BackgroundJobStatusFilter::parse(raw) {
-                Some(filter) => Some(filter),
-                None => {
-                    return ToolExecutionResult::error(json!({
-                        "tool": "jobs_list",
-                        "reason_code": "jobs_invalid_arguments",
-                        "error": "status must be one of queued,running,succeeded,failed,cancelled,terminal",
-                    }))
-                }
-            },
-            None => None,
-        };
-
-        let runtime = match resolve_background_job_runtime(&self.policy) {
-            Ok(runtime) => runtime,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_list",
-                    "reason_code": "jobs_runtime_unavailable",
-                    "error": error,
-                }))
-            }
-        };
-        let jobs = match runtime.list_jobs(limit, status_filter).await {
-            Ok(jobs) => jobs,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_list",
-                    "reason_code": "jobs_runtime_error",
-                    "error": error.to_string(),
-                }))
-            }
-        };
-        let health = runtime.inspect_health().await;
-        ToolExecutionResult::ok(json!({
-            "tool": "jobs_list",
-            "reason_code": "jobs_list_ok",
-            "limit": limit,
-            "status": status,
-            "count": jobs.len(),
-            "jobs": jobs.iter().map(|job| background_job_record_payload(job, false)).collect::<Vec<_>>(),
-            "health": background_job_health_payload(&health),
-        }))
-    }
-}
-
-/// Public struct `JobsStatusTool` used across Tau components.
-pub struct JobsStatusTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl JobsStatusTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for JobsStatusTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "jobs_status".to_string(),
-            description: "Fetch one background job manifest and bounded stdout/stderr previews"
-                .to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "job_id": {
-                        "type": "string",
-                        "description": "Background job id returned by jobs_create"
-                    },
-                    "output_preview_bytes": {
-                        "type": "integer",
-                        "description": format!(
-                            "Optional output preview byte cap (default {}, max {})",
-                            JOBS_OUTPUT_PREVIEW_DEFAULT_BYTES,
-                            JOBS_OUTPUT_PREVIEW_MAX_BYTES
-                        )
-                    }
-                },
-                "required": ["job_id"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        if !self.policy.jobs_enabled {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_status",
-                "reason_code": "jobs_disabled",
-                "error": "background jobs tooling is disabled by policy",
-            }));
-        }
-        let job_id = match required_string(&arguments, "job_id") {
-            Ok(job_id) => job_id.trim().to_string(),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_status",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if job_id.is_empty() {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_status",
-                "reason_code": "jobs_invalid_arguments",
-                "error": "job_id must not be empty",
-            }));
-        }
-        let preview_bytes = match optional_usize(
-            &arguments,
-            "output_preview_bytes",
-            JOBS_OUTPUT_PREVIEW_DEFAULT_BYTES,
-            JOBS_OUTPUT_PREVIEW_MAX_BYTES,
-        ) {
-            Ok(preview_bytes) => preview_bytes,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_status",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-
-        let runtime = match resolve_background_job_runtime(&self.policy) {
-            Ok(runtime) => runtime,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_status",
-                    "reason_code": "jobs_runtime_unavailable",
-                    "error": error,
-                }))
-            }
-        };
-        let Some(record) = (match runtime.get_job(job_id.as_str()).await {
-            Ok(record) => record,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_status",
-                    "reason_code": "jobs_runtime_error",
-                    "error": error.to_string(),
-                }))
-            }
-        }) else {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_status",
-                "reason_code": "job_not_found",
-                "job_id": job_id,
-                "error": "background job id does not exist",
-            }));
-        };
-
-        let stdout_preview =
-            read_output_preview(record.stdout_path.as_path(), preview_bytes).unwrap_or_default();
-        let stderr_preview =
-            read_output_preview(record.stderr_path.as_path(), preview_bytes).unwrap_or_default();
-
-        ToolExecutionResult::ok(json!({
-            "tool": "jobs_status",
-            "reason_code": "jobs_status_ok",
-            "job": background_job_record_payload(&record, true),
-            "stdout_preview": stdout_preview,
-            "stderr_preview": stderr_preview,
-            "output_preview_bytes": preview_bytes,
-        }))
-    }
-}
-
-/// Public struct `JobsCancelTool` used across Tau components.
-pub struct JobsCancelTool {
-    policy: Arc<ToolPolicy>,
-}
-
-impl JobsCancelTool {
-    pub fn new(policy: Arc<ToolPolicy>) -> Self {
-        Self { policy }
-    }
-}
-
-#[async_trait]
-impl AgentTool for JobsCancelTool {
-    fn definition(&self) -> ToolDefinition {
-        ToolDefinition {
-            name: "jobs_cancel".to_string(),
-            description: "Cancel one background job by id".to_string(),
-            parameters: json!({
-                "type": "object",
-                "properties": {
-                    "job_id": {
-                        "type": "string",
-                        "description": "Background job id returned by jobs_create"
-                    }
-                },
-                "required": ["job_id"],
-                "additionalProperties": false
-            }),
-        }
-    }
-
-    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
-        if !self.policy.jobs_enabled {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_cancel",
-                "reason_code": "jobs_disabled",
-                "error": "background jobs tooling is disabled by policy",
-            }));
-        }
-        let job_id = match required_string(&arguments, "job_id") {
-            Ok(job_id) => job_id.trim().to_string(),
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_cancel",
-                    "reason_code": "jobs_invalid_arguments",
-                    "error": error,
-                }))
-            }
-        };
-        if job_id.is_empty() {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_cancel",
-                "reason_code": "jobs_invalid_arguments",
-                "error": "job_id must not be empty",
-            }));
-        }
-
-        if let Some(rbac_result) = evaluate_tool_rbac_gate(
-            self.policy.rbac_principal.as_deref(),
-            "jobs_cancel",
-            self.policy.rbac_policy_path.as_deref(),
-            json!({ "job_id": job_id }),
-        ) {
-            return rbac_result;
-        }
-        if let Some(rate_limit_result) =
-            evaluate_tool_rate_limit_gate(&self.policy, "jobs_cancel", json!({ "job_id": job_id }))
-        {
-            return rate_limit_result;
-        }
-
-        let runtime = match resolve_background_job_runtime(&self.policy) {
-            Ok(runtime) => runtime,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_cancel",
-                    "reason_code": "jobs_runtime_unavailable",
-                    "error": error,
-                }))
-            }
-        };
-
-        let Some(record) = (match runtime.cancel_job(job_id.as_str()).await {
-            Ok(record) => record,
-            Err(error) => {
-                return ToolExecutionResult::error(json!({
-                    "tool": "jobs_cancel",
-                    "reason_code": "jobs_runtime_error",
-                    "error": error.to_string(),
-                }))
-            }
-        }) else {
-            return ToolExecutionResult::error(json!({
-                "tool": "jobs_cancel",
-                "reason_code": "job_not_found",
-                "job_id": job_id,
-                "error": "background job id does not exist",
-            }));
-        };
-        let health = runtime.inspect_health().await;
-        ToolExecutionResult::ok(json!({
-            "tool": "jobs_cancel",
-            "reason_code": "jobs_cancel_ok",
-            "job": background_job_record_payload(&record, false),
-            "health": background_job_health_payload(&health),
         }))
     }
 }
